Implement Authentication
This adds GitHub login/logout logic.
As Django depends strongly on a db, this replicates the django logic
but using a more simple storage which consists of signed cookies.

In this way, we don't need to depend on an external database and
we don't need to keep secrets in a centralised system.

Replicating the django logic, allows us to use the standard and well
known way of accessing the logged-in user using request.user and
protecting views with the login_required decorator.